started refactoring the code
