+++ packages/sdk-core/package.json
@@ -31,6 +31,7 @@
     "@cosmjs/proto-signing": "0.26.5",
     "@cosmjs/stargate": "0.26.5",
     "@cosmjs/tendermint-rpc": "0.26.5",
+    "@cosmjs/utils": "^0.27.1",
     "@desmoslabs/proto": "1.0.0",
     "@walletconnect/client": "^1.6.6",
     "@walletconnect/types": "^1.6.6",
